+++ db/__init__.py
@@ -1,0 +1,4 @@
+# Added: 2025-05-13T17:12:00-04:00 - Database package initialization
+from .model_cache import model_cache_db
+
+__all__ = ['model_cache_db']